# api/openapi.yaml
openapi: 3.1.0
info:
  title: Wuling DevOps API
  version: 0.1.0-stage1
  description: |
    Stage 1 of the Wuling DevOps platform.

    What's implemented:
      - Auth (register, login, /me, personal access tokens, SSH public keys)
      - Orgs and Projects (CRUD)
      - Repositories (CRUD + libgit2-backed read APIs)
      - Git smart HTTP (clone/fetch/push) and Git over SSH (default :2222)
      - Issues (CRUD, labels, comments, assignees)
      - Merge Requests (open/diff/merge with FF/merge-commit/squash, comments, reviews)
      - Wiki (per-project Markdown pages backed by a separate Git repo)
      - Insights (per-project activity, contributors, language stats)

    Planned for later stages (URLs reserved, not yet served):
      - Pipelines

    All errors use the JSON envelope:
      { "error": { "code": "<stable_code>", "message": "<human msg>", "details": {} } }
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/avatar:
    put:
      summary: Upload or replace the caller's avatar
      description: |
        Body is a multipart/form-data with a single `avatar` file field. The
        server decodes the upload (PNG/JPEG/GIF only), centre-crops to a
        square, resizes to 256×256, and re-encodes as PNG. Max body size is
        2 MiB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvatarUploadResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
    delete:
      summary: Remove the caller's uploaded avatar
      responses:
        "204": { description: removed }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
  /api/v1/users/{username}/avatar:
    get:
      summary: Public PNG of a user's avatar
      description: |
        Streams the user's uploaded PNG. Returns 404 when the user has not
        uploaded one — frontends should fall back to a deterministic initials
        tile. Response is cacheable for an hour against the query string in
        User.avatar_url.
      security: []
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
        - in: query
          name: v
          schema: { type: integer }
          description: Cache-buster (Unix seconds). Ignored by the server.
      responses:
        "200":
          description: ok
          content:
            image/png: {}
        "404": { $ref: "#/components/responses/NotFoundError" }

  # ---------------- org members ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
      description: |
        Used by Git CLI only. The "password" is either a personal access token
        (preferred — recognised by the `wlpat_` prefix) or the user's account
        password.

        **Security:** credentials are sent on every request, so basic auth
        MUST be used over HTTPS in production — configure server URLs as
        `https://` and terminate TLS in front of the API. For local
        development against `localhost` plain HTTP is acceptable.

        Operators should prefer personal access tokens (`wlpat_…`) over
        account passwords, and consider OAuth/bearer-token flows for
        non-CLI clients.
...
  responses:
    UnauthorizedError:
      description: missing or invalid credentials
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFoundError:
      description: resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: request validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - already_exists
                - rate_limited
                - unsupported
                - bad_request
                - internal
                - unavailable
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    AvatarUploadResponse:
      type: object
      properties:
        avatar_url:        { type: string }
        avatar_updated_at: { type: string, format: date-time }
...